separated out fuzzers and UTs
