+++ chaoscenter/graphql/server/pkg/environment/handler/handler_test.go
@@ -3,13 +3,8 @@ package handler
 import (
 	"context"
 	"errors"
-	fuzz "github.com/AdaLogics/go-fuzz-headers"
-	"github.com/sirupsen/logrus"
 	"testing"
 	"time"
-
-	"github.com/litmuschaos/litmus/chaoscenter/graphql/server/pkg/database/mongodb"
-	"go.mongodb.org/mongo-driver/mongo"
 
 	"github.com/golang-jwt/jwt"
 	"github.com/google/uuid"
@@ -28,9 +23,9 @@ var (
 	environmentOperator = dbOperationsEnvironment.NewEnvironmentOperator(mongodbMockOperator)
 )
 
-const JwtSecret = "testsecret"
+var JwtSecret = "testsecret"
 
-func getSignedJWT(name string) (string, error) {
+func GetSignedJWT(name string) (string, error) {
 	token := jwt.New(jwt.SigningMethodHS512)
 	claims := token.Claims.(jwt.MapClaims)
 	claims["uid"] = uuid.NewString()
@@ -69,7 +64,7 @@ func TestCreateEnvironment(t *testing.T) {
 			expectedEnv: nil,
 			expectedErr: errors.New("invalid Token"),
 			given: func() string {
-				token, err := getSignedJWT("testUser")
+				token, err := GetSignedJWT("testUser")
 				if err != nil {
 					return token
 				}
@@ -144,7 +139,7 @@ func TestDeleteEnvironment(t *testing.T) {
 			},
 			expectedErr: errors.New("invalid Token"),
 			given: func() string {
-				token, err := getSignedJWT("testUser")
+				token, err := GetSignedJWT("testUser")
 				if err != nil {
 					return token
 				}
@@ -188,114 +183,4 @@ func TestDeleteEnvironment(t *testing.T) {
 			}
 		})
 	}
-}
-
-func FuzzCreateEnvironment(f *testing.F) {
-	utils.Config.JwtSecret = JwtSecret
-	f.Fuzz(func(t *testing.T, data []byte) {
-		fuzzConsumer := fuzz.NewConsumer(data)
-		targetStruct := &struct {
-			input     model.CreateEnvironmentRequest
-			projectID string
-		}{}
-		err := fuzzConsumer.GenerateStruct(targetStruct)
-		if err != nil {
-			return
-		}
-		mongodbMockOperator.On("Create ", mock.Anything, mongodb.EnvironmentCollection, mock.Anything).Return(nil).Once()
-		token, err := getSignedJWT("testUser")
-		if err != nil {
-			logrus.Errorf("Error genrating token %v", err)
-		}
-
-		ctx := context.WithValue(context.Background(), authorization.AuthKey, token)
-		service := handler.NewEnvironmentService(environmentOperator)
-
-		env, err := service.CreateEnvironment(ctx, targetStruct.projectID, &targetStruct.input)
-		if err != nil {
-			t.Errorf("Unexpected error: %v", err)
-		}
-		if env == nil {
-			t.Errorf("Returned environment is nil")
-		}
-	})
-}
-
-func FuzzTestDeleteEnvironment(f *testing.F) {
-	utils.Config.JwtSecret = JwtSecret
-	testCases := []struct {
-		projectID     string
-		environmentID string
-	}{
-		{
-			projectID:     "testProject",
-			environmentID: "testEnvID",
-		},
-	}
-	for _, tc := range testCases {
-		f.Add(tc.projectID, tc.environmentID)
-	}
-
-	f.Fuzz(func(t *testing.T, projectID string, environmentID string) {
-
-		findResult := []interface{}{bson.D{
-			{Key: "environment_id", Value: environmentID},
-			{Key: "project_id", Value: projectID},
-		}}
-		singleResult := mongo.NewSingleResultFromDocument(findResult[0], nil, nil)
-		mongodbMockOperator.On("Get", mock.Anything, mongodb.EnvironmentCollection, mock.Anything).Return(singleResult, nil).Once()
-		mongodbMockOperator.On("UpdateMany", mock.Anything, mongodb.EnvironmentCollection, mock.Anything, mock.Anything, mock.Anything).Return(&mongo.UpdateResult{}, nil).Once()
-		token, err := getSignedJWT("testUser")
-		if err != nil {
-			logrus.Errorf("Error genrating token %v", err)
-		}
-
-		ctx := context.WithValue(context.Background(), authorization.AuthKey, token)
-		service := handler.NewEnvironmentService(environmentOperator)
-
-		env, err := service.DeleteEnvironment(ctx, projectID, environmentID)
-		if err != nil {
-			t.Errorf("Unexpected error: %v", err)
-		}
-
-		if env == "" {
-			t.Errorf("Returned environment is nil")
-		}
-	})
-}
-
-func FuzzTestGetEnvironment(f *testing.F) {
-	utils.Config.JwtSecret = JwtSecret
-	testCases := []struct {
-		projectID     string
-		environmentID string
-	}{
-		{
-			projectID:     "testProject",
-			environmentID: "testEnvID",
-		},
-	}
-	for _, tc := range testCases {
-		f.Add(tc.projectID, tc.environmentID)
-	}
-
-	f.Fuzz(func(t *testing.T, projectID string, environmentID string) {
-
-		findResult := []interface{}{bson.D{
-			{Key: "environment_id", Value: environmentID},
-			{Key: "project_id", Value: projectID},
-		}}
-		singleResult := mongo.NewSingleResultFromDocument(findResult[0], nil, nil)
-		mongodbMockOperator.On("Get", mock.Anything, mongodb.EnvironmentCollection, mock.Anything).Return(singleResult, nil).Once()
-		service := NewEnvironmentService(environmentOperator)
-
-		env, err := service.GetEnvironment(projectID, environmentID)
-		if err != nil {
-			t.Errorf("Unexpected error: %v", err)
-		}
-
-		if env == nil {
-			t.Errorf("Returned environment is nil")
-		}
-	})
 }
