+++ chaoscenter/graphql/server/pkg/environment/handler/testdata/fuzz/FuzzCreateEnvironment/582528ddfad69eb5
@@ -1,0 +1,2 @@
+go test fuzz v1
+[]byte("0")
